pyproject.toml
[build-system]
requires = ["setuptools>=75", "wheel>=0.44"]
build-backend = "setuptools.build_meta"

[project]
name = "tournament-system"
version = "0.1.0"
description = "Flask-based tournament management platform for the Internet Applications lab."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "Flask>=3.0",
    "Flask-SQLAlchemy>=3.1",
    "Flask-Migrate>=4.0",
    "Flask-WTF>=1.2",
    "Flask-Login>=0.6",
    "Flask-Mail>=0.10",
    "Flask-Limiter>=3.5",
    "marshmallow>=3.20",
    "python-dotenv>=1.0",
    "email-validator>=2.1",
    "psycopg[binary]>=3.1",
    "argon2-cffi>=23.1",
    "apispec>=6.6",
    "celery>=5.4",
    "redis>=5.0",
    "alembic>=1.13",
    "gunicorn>=22.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-flask>=1.3",
    "factory-boy>=3.3",
    "faker>=28.0",
    "ruff>=0.6",
    "black>=24.4",
    "pre-commit>=3.8"
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.setuptools.packages.find]
exclude = ["migrations*"]
